requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.1.16
langgraph==0.0.39
openai==1.14.3      # or anthropic, google‑generativeai, etc.
python-dateutil==2.9.0
pydantic==2.7.1
python-dotenv==1.0.1
streamlit==1.35.0
requests==2.32.3
google-api-python-client==2.131.0
google-auth==2.30.0
langchain-openai==0.1.7   # or pip freeze | grep langchain-openai for the exact version you just got
